Modifying Data objects to have invertdata. This should make the invertdata in the Dataset unnecessary, I think...
Operation.apply functions return Data object and (optionally) invertdata.
When the operation is called, it calls apply to get the array and the invertdata and creates a Data object with them.
If the apply function only returns one argument, then invertdata is set to None.
Currently, to conform with invertdata in Dataset class being a dict, this is a dict with names corresponding to operation names. However, I think we should change this to a list to be consistent with how operations are stored.

Modified some Operations to return invertdata:
Fusion returns invertdata for its fused operations as a list
Velocity Operations return pose at start time(s). This was slightly complicated because there are potentially multiple start times due to how data are packed, so it returns {'t0s': t0s, 'x0s': x0s} for each isstart timepoint.
flyllm.Pose returns flyid as invertdata

Added helper function multistart_cumsum for inverting Velocity operations with subsets of frames that are nans

Working on evaluate_fly_llm.py